feature_YJ_modify exception, teamfeed controller

+++ BackEnd/src/main/java/com/ssafy/BackEnd/controller/ProfileController.java
@@ -6,6 +6,8 @@ import com.ssafy.BackEnd.entity.ProfileKeyword;
 import com.ssafy.BackEnd.entity.Request.RequestModifyProfile2;
 import com.ssafy.BackEnd.entity.Response;
 import com.ssafy.BackEnd.entity.TeamKeyword;
+import com.ssafy.BackEnd.exception.CustomException;
+import com.ssafy.BackEnd.exception.ErrorCode;
 import com.ssafy.BackEnd.repository.ProfileKeywordRepository;
 import com.ssafy.BackEnd.repository.UserRepository;
 import com.ssafy.BackEnd.service.HashTagAlgorithm;
@@ -15,6 +17,8 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import javassist.NotFoundException;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -29,6 +33,7 @@ import java.util.*;
 @RequestMapping("/profile")
 @Slf4j
 public class ProfileController {
+    private static final Logger logger = LogManager.getLogger(ProfileController.class);
 
     @Autowired
     private ProfileService profileService;
@@ -55,10 +60,12 @@ public class ProfileController {
             Optional<Profile> profile = profileService.findById(profile_id);
             if (profile != null) {
                 System.out.println("profile : "+profile.get().getEmail());
+                logger.info("INFO SUCCESS");
                 return new ResponseEntity<Profile>(profile.get(), HttpStatus.OK);
             }
             else {
-                return new ResponseEntity<Profile>((Profile) null, HttpStatus.NOT_FOUND);
+                throw new CustomException(ErrorCode.NO_DATA_ERROR);
+                //return new ResponseEntity<Profile>((Profile) null, HttpStatus.NOT_FOUND);
             }
 
         } catch (Exception e) {
@@ -71,33 +78,47 @@ public class ProfileController {
     public ResponseEntity<Profile> modifyProfile(@PathVariable Long profile_id, @RequestBody RequestModifyProfile2 requestModifyProfile2) {
         Response response = new Response();
         try {
-
             Profile findProfile = profileService.findById(profile_id).get();
             Profile newProfile = profileService.modifyProfile(findProfile, requestModifyProfile2);
             log.info("수정완료");
+            logger.info("INFO SUCCESS");
             return new ResponseEntity<Profile>(newProfile, HttpStatus.OK);
         } catch (Exception e) {
             log.info(e.toString());
-            return new ResponseEntity<Profile>((Profile) null, HttpStatus.NOT_FOUND);
+            throw new CustomException(ErrorCode.NO_DATA_ERROR);
+            //return new ResponseEntity<Profile>((Profile) null, HttpStatus.NOT_FOUND);
         }
     }
 
     @PostMapping("/update/image/{profile_id}")
     public ResponseEntity<String> updateImage(@PathVariable Long profile_id, MultipartFile multipartFile) throws NotFoundException {
         String imagePath = imageService.update(profile_id, multipartFile);
+        if (imagePath == null) {
+            logger.info("NO IMAGE PATH");
+            //throw new CustomException(ErrorCode.NO_DATA_ERROR);
+        }
+        logger.info("INFO SUCCESS");
+
         return new ResponseEntity<String>(imagePath, HttpStatus.OK);
     }
 
     @DeleteMapping("/{profile_id}")
     public void deleteUser(@PathVariable Long profile_id) throws NotFoundException {
-
         Optional<Profile> findProfile = profileService.findById(profile_id);
+        if (findProfile == null) {
+            throw new CustomException(ErrorCode.NO_DATA_ERROR);
+        }
         profileService.deleteUser(profile_id);
+        logger.info("INFO SUCCESS");
     }
 
     @PostMapping("/keyword/{profile_id}")
     public ResponseEntity<List<ProfileKeyword>> addKeyword(@PathVariable Long profile_id, @RequestParam String content) throws NotFoundException {
         Profile profile = profileService.findById(profile_id).get();
+        if (profile == null) {
+            throw new CustomException("no profile", ErrorCode.NO_DATA_ERROR);
+        }
+
         List<String> keywords = hashTagAlgorithm.strList(content);
         List<ProfileKeyword> profileKeywords = profile.getProfile_keyword();
         for (String keyword : keywords) {
@@ -115,6 +136,7 @@ public class ProfileController {
             }
         }
         profile.setProfile_keyword(profileKeywords);
+        logger.info("INFO SUCCESS");
         return new ResponseEntity<List<ProfileKeyword>>(profileKeywords, HttpStatus.OK);
 
     }
@@ -122,6 +144,10 @@ public class ProfileController {
     @GetMapping("/keyword/{profile_id}")
     public List<ProfileKeyword> getKeywords(@PathVariable Long profile_id) throws NotFoundException {
         Profile profile = profileService.findById(profile_id).get();
+        if (profile == null) {
+            throw new CustomException("no profile", ErrorCode.NO_DATA_ERROR);
+        }
+
         List<ProfileKeyword> profileKeywords = profileService.getProfileKeywords(profile_id);
         List<TeamKeyword> teamKeywords = profileService.getTeamKeywords(profile_id);
         List<ProfileKeyword> finalKeywords = new ArrayList<>();
@@ -159,6 +185,7 @@ public class ProfileController {
             System.out.println(profileKeyword.getCount());
         }
 
+        logger.info("INFO SUCCESS");
         return finalKeywords;
     }
 }
